backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-multipart
PyJWT
passlib
bcrypt
pydantic>=2.7,<3.0
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-groq>=0.2.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.4.0
groq
PyPDF2
python-docx
duckduckgo-search
python-dotenv
